requirements.txt
flask
requests
openai-whisper
gunicorn
